feat: create basic html of createroompage
